refactor: 체스판의 행 자료구조를 List<Piece>에서 Rank로 변경
